Student-Assistance-main/student-support-backend/requirements.txt
flask
flask-cors
tensorflow
numpy
scikit-learn
langdetect
deep-translator
python-dotenv
pymongo
bcrypt
vaderSentiment
gunicorn
